Lazy-load spec files
When you have a really large app and test suite, it can take a long time
for RSpec to boot up. In the case of canvas-lms, it's two minutes before
test-queue can even start. That is no bueno, but this commit fixes that.

Reach into the guts of RSpec and twist it all around so that we don't
ever require the spec files in the master process, and only require
assigned ones in each worker. On the master, files get loaded in a separate
process so that we don't block anything. It reports back to the main process
so it has the info it needs for the queues, and then work can be assigned
out and started virtually immediately.